confirm page fix and sin-utils addition
